refactor: <김상현 #7> refactoring directory structure
