+++ src/routes/login/+page.svelte
@@ -1,0 +1,19 @@
+<script>
+
+</script>
+
+<form>
+    <h1>Login</h1>
+
+    <input type="email" name="email"/>
+    <input type="password" name="password"/>
+
+    <button>Login</button>
+</form>
+
+<style>
+    :global(.touched:invalid) {
+        border-color: red;
+        outline-color: red;
+    }
+</style>
