+++ src/routes/sidebar.svelte
@@ -8,18 +8,18 @@
     <div class="d-flex flex-column align-items-center align-items-sm-start px-3 pt-2 text-white min-vh-100"
         style="margin-left:-20px">
         <ul class="nav nav-pills flex-column mb-sm-auto mb-0 align-items-center align-items-sm-start" id="menu">
-            <li class="nav-item " class:active-menu={$page.url.pathname=='/'}>
-                <a href="/" class="nav-link ms-4 align-middle px-0 fs-4 ">
+            <li class="nav-item " class:active-menu={$page.url.pathname=='/management/reserve'}>
+                <a href="/management/reserve" class="nav-link ms-4 align-middle px-0 fs-4 ">
                     <Fa  icon={faCalendarCheck} />  <span class="ms-1 text-black-50 d-none d-sm-inline">예약</span>
                 </a>
             </li>
-            <li class="nav-item" class:active-menu={$page.url.pathname=='/match'}>
-                <a href="/match" class="nav-link align-middle fs-4 px-0">
+            <li class="nav-item" class:active-menu={$page.url.pathname=='/management/match'}>
+                <a href="/management/match" class="nav-link align-middle fs-4 px-0">
                     <Fa class="text-light" icon={faCircle} /> <span class="ms-1 text-black-50 d-none d-sm-inline">매칭</span>
                 </a>
             </li>
-            <li class="nav-item" class:active-menu={$page.url.pathname=='/seat'}>
-                <a href="/seat" class="nav-link align-middle fs-4 px-0">
+            <li class="nav-item" class:active-menu={$page.url.pathname=='/management/seat'}>
+                <a href="/management/seat" class="nav-link align-middle fs-4 px-0">
                     <Fa class="text-light" icon={faCircle} /> <span class="ms-1 text-black-50 d-none d-sm-inline">좌석</span>
                 </a>
             </li>
